[change]ファイルの割り振り
